@@ -1,6 +1,6 @@
 {
   "name": "monoteamf",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -12,43 +12,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.2",
-    "@angular/common": "^5.0.2",
-    "@angular/compiler": "^5.0.2",
-    "@angular/core": "^5.0.2",
-    "@angular/forms": "^5.0.2",
-    "@angular/http": "^5.0.2",
-    "@angular/platform-browser": "^5.0.2",
-    "@angular/platform-browser-dynamic": "^5.0.2",
-    "@angular/router": "^5.0.2",
-    "angular2-flash-messages": "^1.0.8",
+    "@angular/animations": "^5.1.2",
+    "@angular/common": "^5.1.2",
+    "@angular/compiler": "^5.1.2",
+    "@angular/core": "^5.1.2",
+    "@angular/forms": "^5.1.2",
+    "@angular/http": "^5.1.2",
+    "@angular/platform-browser": "^5.1.2",
+    "@angular/platform-browser-dynamic": "^5.1.2",
+    "@angular/router": "^5.1.2",
+    "angular2-flash-messages": "^2.0.4",
     "angular2-jwt": "^0.2.3",
-    "bulma-checkradio": "^0.1.0",
-    "core-js": "^2.5.1",
-    "jwt-decode": "^2.2.0",
-    "rxjs": "^5.5.2",
+    "core-js": "^2.5.3",
+    "rxjs": "^5.5.6",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.2",
-    "@angular/compiler-cli": "^5.0.2",
-    "@angular/language-service": "^5.0.2",
+    "@angular/cli": "^1.6.2",
+    "@angular/compiler-cli": "^5.1.2",
+    "@angular/language-service": "^5.1.2",
     "@types/jasmine": "~2.8.2",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.0.53",
-    "codelyzer": "~4.0.1",
+    "@types/node": "^8.5.2",
+    "codelyzer": "^4.0.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^1.7.1",
+    "karma": "^2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.2.0",
-    "ts-node": "~3.3.0",
+    "protractor": "^5.2.2",
+    "ts-node": "~4.0.2",
     "tslint": "~5.8.0",
-    "typescript": "~2.6.1",
-    "webpack": "^3.8.1"
+    "typescript": "^2.6.2",
+    "webpack": "^3.10.0"
   }
 }
