@@ -1,12 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { ReactiveFormsModule } from '@angular/forms';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { JwtHelper } from 'angular2-jwt';
+import { FlashMessagesModule } from 'angular2-flash-messages';
 
 import { AppComponent } from './app.component';
+
+import { DashboardComponent } from './dashboards/components/dashboard.component';
 
 import { LoginComponent } from './accounts/login/components/login.component';
 
@@ -20,48 +22,54 @@ import { CharacterSidebarComponent } from './lobbies/components/character-sideba
 import { ListLobbyComponent } from './lobbies/components/list-lobby.component';
 import { NewLobbyComponent } from './lobbies/components/new-lobby.component';
 import { LobbyComponent } from './lobbies/components/lobby.component';
-import { DisplayComponentService } from './lobbies/display-component.service';
 
 import { GroupComponent } from './groups/components/group.component';
 
-import { NavbarComponent } from './shared/components/navbar/navbar.component';
-import { FooterComponent } from './shared/components/footer/footer.component';
+import { AuthGuardService as AuthGuard } from './services/guards/auth-guard.service';
+import { GroupGuardService as GroupGuard } from './services/guards/group-guard.service';
+import { HomePageGuardService as HomePageGuard } from './services/guards/home-page-guard.service'
+
+import { TokenInterceptor } from './services/interceptors/token.interceptor';
 
 import { AuthService } from './services/auth.service';
 import { CharacterService } from './services/character.service';
 import { RegisterService } from './services/register.service';
-import { AuthGuardService as AuthGuard } from './services/auth-guard.service';
-import { GroupGuardService as GroupGuard } from './services/group-guard.service';
 import { LobbyService } from './services/lobby.service';
+import { UserService } from './services/user.service';
 
-import { TokenInterceptor } from './services/interceptors/token.interceptor';
+import { NavbarComponent } from './shared/components/navbar/navbar.component';
+import { FooterComponent } from './shared/components/footer/footer.component';
+
+import { TypefightString } from './shared/pipes/typefight-string.pipe';
 
 const appRoutes: Routes = [
-  {path: '', component:LoginComponent},
+  {path: '', component:LoginComponent, canActivate: [HomePageGuard]},
   {path: 'register', component:RegisterComponent},
+  {path: 'dashboard', component:DashboardComponent, canActivate: [AuthGuard]},
   {path: 'players/:id/characters', component:CharacterComponent, canActivate: [AuthGuard]},
   {path: 'players/:id/search', component:LobbyComponent, canActivate: [AuthGuard, GroupGuard]},
-  {path: 'groups/:id', component:GroupComponent, canActivate: [AuthGuard]}
+  {path: 'players/:id/groups/:id', component:GroupComponent, canActivate: [AuthGuard]}
 ]
 
 @NgModule({
   declarations: [
     AppComponent,
     NavbarComponent, FooterComponent,
+    DashboardComponent,
     LoginComponent,
     RegisterComponent,
     CharacterComponent, CharactersListComponent, NewCharacterComponent,
     LobbyComponent, CharacterSidebarComponent, ListLobbyComponent, NewLobbyComponent,
-    GroupComponent
+    GroupComponent,
+    TypefightString
   ],
   imports: [
-    BrowserModule, FormsModule, ReactiveFormsModule,
-    HttpClientModule, RouterModule.forRoot(appRoutes)
+    BrowserModule, ReactiveFormsModule, HttpClientModule,
+    FlashMessagesModule.forRoot(), RouterModule.forRoot(appRoutes)
   ],
   providers: [
-    AuthService, RegisterService, CharacterService, LobbyService,
-    DisplayComponentService,
-    AuthGuard, GroupGuard,
+    AuthService, RegisterService, CharacterService, LobbyService, UserService,
+    AuthGuard, GroupGuard, HomePageGuard,
     {provide: HTTP_INTERCEPTORS, useClass: TokenInterceptor, multi: true},
     JwtHelper
   ],
