@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Router } from '@angular/router';
-import { Observable } from 'rxjs';
-import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs/Observable';
 import { tokenNotExpired } from 'angular2-jwt';
 
 @Injectable()
