@@ -1,8 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import * as JWT from 'jwt-decode';
 
+import { JwtHelper } from 'angular2-jwt';
 import { CharacterService } from '../../services/character.service';
+import { FlashMessagesService } from 'angular2-flash-messages';
+
+import { Character } from '../../models/character';
 
 @Component({
   selector: 'app-new-characters',
@@ -12,41 +15,40 @@ import { CharacterService } from '../../services/character.service';
 export class NewCharacterComponent implements OnInit {
   newCharacterForm: FormGroup;
   error: string = null;
+  character: Character;
 
   servers: string[] = [
     'Agride', 'Beta test', 'Djaul', 'Écho', 'Goultard', 'Hel Munster', 'Illyazelle',
     'Julith', 'Mériana', 'Merkator', 'Mylaise', 'Nidas', 'Pandore', 'Ush'
   ];
+
   categories: string[] = [
     'Crâ', 'Ecaflip', 'Eliotrope', 'Eniripsa', 'Enutrof', 'Féca', 'Huppermage', 'Iop', 'Osamodas',
     'Ouginak', 'Pandawa', 'Roublard', 'Sacrieur', 'Sadida','Sram', 'Steamer', 'Xélor', 'Zobal'
   ];
-  roles: string[] = [
-  ];
 
-  constructor(private fb: FormBuilder, private characterService: CharacterService) {}
+  constructor(private jwtHelper: JwtHelper, private fb: FormBuilder, private characterService: CharacterService, private flashMessagesService: FlashMessagesService) {}
 
   ngOnInit() {
-    //Generate form
     this.newCharacterForm = this.fb.group({
-      pseudo: ['', Validators.required],
+      pseudo: ['', [Validators.required, Validators.minLength(3), Validators.maxLength(33)]],
       server: ['', Validators.required],
-      role: [''],
       category: ['', Validators.required],
       level: ['', Validators.required]
     });
   }
 
   onNewCharacterSubmit(): void {
-    let auth_token = localStorage.getItem('auth_token');
-    let currentUserId = JWT(auth_token)['sub'];
-
-    let character = this.newCharacterForm.value;
-    character.user_id = currentUserId;
-
-    this.characterService.newCharacter(character).subscribe(
+    const userId: number = this.jwtHelper.decodeToken(localStorage.getItem('auth_token'))['sub']
+    let characterForm = this.newCharacterForm.value;
+    characterForm.user_id = userId;
+    this.characterService.createCharacter(characterForm).subscribe(
+      data => {
+        this.character = data;
+        this.characterService.addCharacter(this.character);
+      },
       () => {
-        this.characterService.addCharacter(character);
+        this.flashMessagesService.show('Nouveau personnage ajouté', {cssClass: 'notification is-success'})
         this.newCharacterForm.reset();
       }
     );
