@@ -1,8 +1,8 @@
 export interface Character {
   id?: number;
+  user_id?: number;
   pseudo?: string;
   server?: string;
   category?: string;
-  role?: string;
   level?: number;
 }
