@@ -5,13 +5,11 @@ import { JwtHelper } from 'angular2-jwt';
 
 @Injectable()
 export class GroupGuardService implements CanActivate {
-
   constructor(private jwtHelper: JwtHelper, private router: Router) {}
 
   canActivate(): boolean {
-    const token = localStorage.getItem('auth_token');
-    const userId = this.jwtHelper.decodeToken(token)['sub'];
-    const isOngoingGroup = this.jwtHelper.decodeToken(token)['ongoing_g'];
+    const userId = this.jwtHelper.decodeToken(localStorage.getItem('auth_token'))['sub'];
+    const isOngoingGroup = this.jwtHelper.decodeToken(localStorage.getItem('auth_token'))['ongoing_g'];
     if (isOngoingGroup) {
       this.router.navigate([`/players/${userId}/characters`]);
       return false;
