@@ -1,16 +1,16 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate } from '@angular/router';
 
-import { AuthService } from './auth.service';
+import { AuthService } from './../auth.service';
 
 @Injectable()
-export class AuthGuardService implements CanActivate {
+export class HomePageGuardService implements CanActivate {
 
   constructor(public auth: AuthService, private router: Router) {}
 
   canActivate(): boolean {
-    if (!this.auth.isAuthenticated()) {
-      this.router.navigate(['/']);
+    if (this.auth.isAuthenticated()) {
+      this.router.navigate(['dashboard']);
       return false;
     }
     return true;
